moving sample responses to _utils.js
